Changes to filestructure
