--- src/app/react/src/components/screens/HomePage.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-
-function HomePage() {
-  return (<div>
-  </div>);
-}
-
-export default HomePage;
