+++ src/app/react/src/components/sidebars/ProjectSidebar/index.jsx
