+++ src/app/react/src/components/sidebars/TabbedSidebar/index.jsx
